Refactor getting and setting of config options code.
Config options are now stored in AWSShell attributes and loaded on startup.  On shutdown, the configs are set based on the AWSShell attributes.  The KeyManager class expects a callable to get/set the config options: For the get methods we simply use a lambda that wraps the AWSShell atribute; for the set methods we're forced to define actual methods.  Also moves the call to AWSShell.stop_input_and_refresh_cli to within the key handlers of KeyManager.  In the future we might want to wrap the calls to self.model_completer.match_fuzzy with a convenience function. Updates tests and docstrings.