Extract dot commands to separate object
This extracts out the handling of dot commands to
a separate object.  This makes it easier to add dot
commands, and makes testing dot commands easier.

As part of this work, I've also updated the `.edit`
command to support the EDITOR env var, and fall back
to platform specific defaults if the var is not defined.

Fixes #43
Fixes #40