Implement #29: Maintain history between sessions.
We now use FileHistory as our general purpose history which is persisted between sessions.  We also use an InMemoryHistory so that the special .edit command will only show commands from the current session.  In the future, it might be cleaner to combine these two history attributes into a History class.