Change ctrl-c to do a keyboard interrupt vs app exit.
It seems to be the more standard behavior to use ctrl-c to keyboard interrupt and ctrl-d to exit the app.  This behavior is consistent with the python interactive shell. Making this change allows users to quickly cancel out of the current command.  Users can still use ctrl-d or f10 to exit the app.